# email: remove `json-stringify-safe` to boost perf