--- a/lib/email.js
+++ b/lib/email.js
@@ -1,48 +1,50 @@
 'use strict';
 
 var utilis = require('core-util-is'),
-    querystring = require('querystring'),
-    stringify = require('json-stringify-safe');
+    querystring = require('querystring');
 
 var EmailPattern = RegExp(/[a-z0-9_+&*-]+(?:\.[a-z0-9_+&*-]+)*@(?:[a-z0-9-]+\.)+[a-z]{2,7}/ig);
 
-function maskEmail(data) {
-    var datastr = data,
-        isObject = false,
-        mightHaveReplaced = false;
+function maskCircular(data, seen) {
+  if (utilis.isString(data)) {
+      if (data.indexOf('@') !== -1 || data.indexOf('%40') !== -1) {
+          var datastr = querystring.unescape(data);
+          datastr = datastr.replace(EmailPattern, 'xxx@xxx.xxx');
+          return datastr;
+      }
+      return data;
+  } else if (!utilis.isObject(data)) {
+      return data;
+  }
 
-    if (utilis.isObject(data)) {
-        datastr = stringify(data);
-        isObject = true;
-    }
+  if (seen === null) {
+      seen = [ data ];
+  } else {
+      if (seen.indexOf(data) !== -1)
+          return '[Circular]';
+      seen.push(data);
+  }
 
-    //only run regex if dataString contains @ or %40 (encoded @)
-    if (utilis.isString(datastr) && (datastr.indexOf('@') !== -1 || datastr.indexOf('%40') !== -1)) {
-        datastr = querystring.unescape(datastr);
-        datastr = datastr.replace(EmailPattern, 'xxx@xxx.xxx');
-        mightHaveReplaced = true;
-    }
+  if (Array.isArray(data)) {
+      var copy = new Array(data.length);
+      for (var i = 0; i < data.length; i++)
+          copy[i] = maskCircular(data[i], seen);
+      return copy;
+  }
 
-    if (isObject) {
-        if (mightHaveReplaced) {
-            datastr = parseJSON(datastr);
-        } else {
-            datastr = data;
-        }
+  var copy = {};
+  var keys = Object.keys(data);
+  for (var i = 0; i < keys.length; i++) {
+      var key = keys[i];
+      var value = data[key];
 
-    }
-
-    return datastr;
+      copy[key] = maskCircular(value, seen);
+  }
+  return copy;
 }
 
-function parseJSON(value) {
-    var obj;
-    try {
-        obj = JSON.parse(value);
-    } catch (e) {
-        //No op
-    }
-    return obj;
+function maskEmail(data) {
+    return maskCircular(data, null);
 }