Boost performance by walking through objects manually.

Ops/sec before the patch:

    maskEmail (object) x 943,982 ops/sec ±0.52% (96 runs sampled)
    maskEmail (hit) x 1,063,417 ops/sec ±0.63% (90 runs sampled)
    maskEmail (miss) x 12,863,284 ops/sec ±0.96% (92 runs sampled)

After the patch:

    maskEmail (object) x 5,103,084 ops/sec ±1.75% (86 runs sampled)
    maskEmail (hit) x 1,068,074 ops/sec ±0.59% (92 runs sampled)
    maskEmail (miss) x 12,219,161 ops/sec ±0.80% (92 runs sampled)

--------

This probably needs a major version bump. The `[Circular]` is going to be used for all circular objects, regardless of their position in the original object.