@indutny 2017-06-11
cc @skoranga just in case