@skoranga 2017-06-12
#4 
published `maskme@1.0.0`